+++ app/src/main/java/com/example/daggerhilttest/models/v2/HourlyForecastDataV2.kt
@@ -1,7 +1,7 @@
 package com.example.daggerhilttest.models.v2
 
 data class HourlyForecastDataV2(
-    val temp: Double,
+    val temp: Float,
     val formattedTime: String,
     val icon: String
 )
