+++ app/src/main/java/com/example/daggerhilttest/models/v2/HourlyForecastResponseV2.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class HourlyForecastResponseV2(
     @SerializedName("temp")
-    val temp: Double?,
+    val temp: Float?,
     @SerializedName("datetimeEpoch")
     val dateTimeEpoch: Long?,
     @SerializedName("icon")
